Add a MonadBaseControl instance for PropertyT (#328)
Under an assumption I believe is reasonable (see [this issue](basvandijk/monad-control#48) and [this SO question](https://stackoverflow.com/questions/58105759/monadbasecontrol-laws) I believe it is possible to write a lawful `MonadBaseControl` instance for `CofreeT`, and therefore also for the `TreeT`, `GenT`, and `PropertyT` types in this package. I can't say for sure whether it "makes sense", but hey, maybe it does.