Make `interleave` logarithmtic rather than linear (#313)
The version of `interleave` before this PR was closely modelled on my
blog post; a tiny bit _too_ closely; it wasn't the intention of the blog
post to define "good" shrinkers. I had a footnote for that but should
have made it a bit more prominent. Anyway, in this PR I change
`interleave` from being linear (removing one element at a time) to
logarithmic (trying to remove as big a chunk as possible at a time,
trying the whole list, half the list, etc.), matching the behaviour of
QuickCheck's `shrinkList`. This can have rather dramatic consequences
for the performance of shrinking.