Don't swallow errors if we can't find the source file
There are two possible ways which a source can't be printed at a
location; the location isn't known, or we can't find the source.

In either of these we should fall back to reporting it simply,
while previously, if the source file couldn't be found we would
simply elide the error message.