Fix "reproduce" message.
The previous message would suggest

    recheck (size) (seed) $ withSkip "(skip)" <property>

but this didn't work. `recheck` takes the size and seed of the failed
test, but `withSkip` assumes you're replaying from a given initial seed
at size 0.

We now suggest

    recheckAt (seed) "(skip)" <property>

where the seed is now the initial seed, not the seed that eventually
failed.

This also means that if you see that message, you can rerun with

    HEDGEHOG_SEED='...'

or

    HEDGEHOG_SEED='...' HEDGEHOG_SHRINK='...'

to reproduce the failure, instead of having to edit the source code.
Previously this wouldn't work; `HEDGEHOG_SEED` has to take the initial
seed, which was completely lost.

Since `FailureReport` no longer has a seed, it's now impossible to
generate a `recheck` line that will work, with or without `withSkip`.
But we no longer suggest `recheck` to users.