Fix overflow issue for exponential ranges.
Because Doubles are used in the calculation of the exponential
function, if the word size is greater than 53 then the Double
representation may be inexact and might end up out of the bounds.

The simplest fix is to special case on the size parameter when
calculating the bound, this allows us to just return the bound
and skip any conversions.